requirements.txt
crewai
crewai_tools
python-dotenv
langchain
fastapi
uvicorn
pydantic